Bind free variables in clang-w-fodder snippets.
These changes make it possible to insert code snippets into a
genome from the JSON database while ensuring that any free variables
get bound to something in scope (maybe of the wrong type, but..).
Variable bindings are chosen by ordering the declared variables
with the closest definition first and choosing randomly with an
exponentially decaying probability distribution, so nearby
definitions are chosen preferentially. The code snippets in the
database do not include trailing semicolons, so during
clang-mutate's -set-value we see if the statement about to be
overwritten has a trailing semicolon; if so, we add a semicolon
to the snippet, too.