Add type declarations in clang-mito.
clang-mito gets a new add-type method that takes a hash value of
the type to add. More details are found by looking up the hash in
the global *type-database*.  Additional types that are required are
also added transitively during add-type.  The types used by a clang-mito
instance are kept as a set, and also as a list sorted in reverse
dependency order so that we don't have to do a topological sort on the
types every time we generate the clang-mito's genome.

clang-w-fodder's prepare-code-snippet method now adds the appropriate types
from the snippet into the instance's clang-mito.