Change some uses of genome to genome-string.
This genome-string vs genome thing seems confusing. The problem is
that the clang genome is a combination of the "raw" genome
(program text) with the "mitochondrial" genome (include files,
macro declaractions, types, etc). The first one consists of stuff
that clang-mutate can manipulate nicely, the second consists of
things required to get the first to compile.  So lots of
operations on the clang raw genome are of the form "merge the
clang-mito genome with the raw clang genome, do some processing,
and extract the raw genome back from the result."  This can potentially
change things like the AST numbering, leading to out-of-bounds errors
and other nastiness if you accidentally do something with
(genome clang) rather than (genome-string clang), or vice versa.