Add two crossover strategies and improve insert.
This adds two crossover methods that attempt to perform crossover
that (1) rebinds free variables in the new context of a snippet,
and (2) respects the change in scope depth from the start to the
end of the snippet.

The crossover method selects between two crossover strategies:
single statement crossover, and 2-point-outward. The first selects a
single full statement from each parent and swaps them. The second
is a little more complex to explain.  It first selects a full statement
in one parent, and computes the "across-and-out" successors of the statement.
e.g.

int main() {
  int x = 0;
  for (int i = 0; i < 10; ++i) {
     x += i; // <== Crossover selected this point
     if (x < 0) {
       return 0;
     }
     printf("%d\n", i);
  }
  return x;
}

The across-and-out transitive successors of the selected statement are
[Depth 0]:
  (0.1) x += i;
  (0.2) if (x < 0) { return 0; }
  (0.3) printf("%d\n", i);
[Depth 1]:
  (1.1) return x;

The second crossover point will be taken to be one of these statements.
Two crossover points are then selected in b, ensuring that the depth
of the second crossover point in b matches that in a.  This ensures that
when the snippets are crossed, the result has the correct nesting of
scopes.

Variables are rebound in such a way as to respect the total scope
nesting depth. In the above example, "x" cannot be bound to a
variable defined in the innermost scope at the insertion point, since
it is referenced at depth 1 in the snippet.

In addition to the crossover changes, this modifies the :insert mutation
on clang to rebind free variables in the inserted statement.