Fixing caching of compilation results.
Previously, we never actually compiled anything as the AST phenome superclass
method was never called.  Additionally, we need to cache the original source
file name used when compiling the binary as it is what is present in the
DWARF information.  The original file name is required to create a mapping
of source -> bytes.  Finally, fixing a number of small errors where an
unexpected nil could be returned.