Randomly allow binding free variables to globals.
Previously, it was up to the caller to decide if free variables should
be allowed to be bound to globals.  Now, there is a random chance that
we consider globals when looking for a binding.  If the caller wants to
always allow or disallow global binding, they can just override this
probability in the lexical environment.