Cache JSON DBs as stored CL objects.
When loading a JSON database foo.json, this looks for a file in
the same directory with the name foo.dbcache. If it does not exist
or foo.dbcache's timestamp is older than foo.json's, then foo.json
is loaded like usual and then cached to foo.dbcache.  Otherwise,
foo.json is ignored and the database is directly de-serialized
from foo.dbcache.