Replacing JSON-file databases with Mongo.
clang-w-fodder software objects now utilize an external Mongo database
for fodder selection.  The global variable *mongo-connection* contains
the database connection information.  The find-snippets method can be
utilized to retrieve random snippets from the database.

As part of this change, pick-json was renamed to pick-snippet to recognize
the action of the method instead of the type of the underlying data store.