Adding support for JSON-file and Mongo databases.
All databases now implement the fodder database interface
defined in software-evolution:software/fodder-database.lisp, and thus,
can be queried in an implementation agnostic mannger.

For JSON-file database, the contents of the database are that of the
file specified.

For Mongo databases, we parse a JSON configuration file for the Mongo
database name, host, and port.  The following is an example configuration:
{ "mongo-db": <db_name>[, "host": <host>, "port": <port>] }