Misc. updates - see detail.
clang-mito now implements the find-type interface.  clang-mito instances
can now be utilized as parameters to add-types as type databases.  This
is useful for crossovers of two software objects where we need to add
the types of one software object to another.

update-mito-from-snippet now takes a type-database parameter.  This
removes the need for an add-types-for-snippet method.