Compare disassembly line by line.
The diff function has an expected running time of O(N + PD)
and a worst case of O(NP) where N is the number of elements in the
longest sequence B, D is the shortest edit script between two
sequence A and B, and P is the number of deletions in such a script.

By limiting N, we can increase performance.