Rework analyze-mutation.
new-individual now returns the new variant and a list of the mutation
info. analyze-mutation expects the same as arguments and destructures
the mutation info to get parents, crossover points, etc. This simplifies
all the intermediate code which has to pass those values around, and
will make it easy to change the mutation info if we need to.

generational-evolve calls analyze-mutation after the population is
evaluated. But it still needs a TEST argument to pass through, because
the mutation info can contain intermediate objects that never made it
into the population.