Support for constant adaptive mutation.
Two new generic methods, mutation-types-clang and mutate-clang have been added.

mutation-types-clang should return a list of pairs (mutation-type . probability)
with the mutation types and relative probabilities of each mutation supported.
Subclasses of clang should override this as appopriate for additional new mutation
types.

mutate-clang follows the general contract of mutate, but takes an additional
parameter, mutation-type, which guides the type of mutation to be performed.