Reverting 72bc8c5.
The clang-mutate executable will sometimes produce
JSON output when there are compiler errors (exit 1).
This output is not guaranteed to respect several invariants.
For instance, we assume the root of each AST is a compound
statement; this may not hold on non-zero exits.
Because of this, we cannot use the JSON output on a non-zero
exit without the risk of errors in later processing.