Fix mongo-database:find-snippets.
There were multiple problems with the previous implementation:
1) We could stuck in infinite recursion if limit was small (e.g. 1).  In this case,
in each recursive call, expanded-window is set to 2.
2) By solving (1), the implementation would still be inefficient for small values
of limit as we would have to expand the window to 2, 4, 8, 16, 32, 64, 128, etc.
until we found a snippet which satified the class/full-stmt predicates.
3) The solution was biased against snippets close to zero. Instead of
(kv ($>= "random" window-base) ($<= "random" (+ window-base frac))) we needed
(kv ($>= "random" (- window-base (/ frac 2)))
    ($<= "random" (+ window-base (/ frac 2))))
4) The solution did not sort snippets by the "random" field prior executing a query.
This biased snippets 'near the beginning of the database' as these would match
the query first.

The new solution requires at most (2) queries and is completely randomized.