Include all possible scope vars in bind-free-vars
Previously, bind-free-vars was removing some in-scope vars from the
list of potential variables to rebind to based on the level in the
snippet where the free variable was bound. Because snippets may not
have any relation to the point in the clang object where they're
being inserted, this removal doesn't make sense and may prevent
bind-free-vars from succeeding in some cases (e.g., test case) when
it should succeed.