Make update-asts also update headers
Also make update-headers use push-include and push-macro to add macros
and includes to software object fields without updating the genome.
When fields are being initialized, the includes and macros are already
in the genome, so they shouldn't be added again. (But note that
update-headers-from-snippet may be adding new includes/macros so should
still update the genome).

Additionally removed an incorrect comment from add-include: updating
genome-string also updates genome and nullifies the asts.