Fixup for index OOB exception in bind-free-vars.
When a snippet is inserted, compilation errors/warnings may
occur after adding types/macros required for a snippet. This
could decrease the number of ASTs in the software object. In
this case, we should guard against attempting to insert the
snippet in a location which is no longer valid.

We could be more proactive and have more complex logic to ensure
we always insert a compiling snippet.  However, I feel the cost
of this logic would outway the benefits for this (relatively rare)
situation.  We should allow the non-compiling individual to be
filtered out from the population and move on.

This commit also moves logic for updating the includes/types/macros
for a software object from the fodder selection method into
the recontextualization routine.