Corrections per eschulte peer review.
Note: This was commit ac5e0c6c on the buildbot-2016.10.14 branch.  It was not
merged into eschulte-2016-10-17 which eventually became master.

1) Renamed empty -> database-emptyp.  Note: (defgeneric emptyp ... )
yields a warning "Generic function emptyp clobbers earlier ftype proclamation"
due to a ftype declaration in alexandria with the same name.  This name
collision causes problems as emptyp is utilized by the elf repo
(elf/utils.lisp:51/53).
2) Added comment about limitations of sb-ext:finalize.
3) Removed +pliny-db-host+/+pliny-db-port+ constants in favor of the literal
value appearing inline in the definition of pliny-database.
4) Removed +server-frontend-num-threads+/+server-backend-num-threads+/
+memory-per-thread+ constants in favor of fields on the pliny-database object.
5) Made "GTServer"/"GTServerShutdown"/"GTLoader"/"GTQuery" command names
fields on the pliny-database object.