Remove memoization from json-fodder-database.
This was causing frankenstein-can-rebuild-original unit test failures in bed.

The memoize function, as currently written, computes a hash based on the arguments
to the function to determine if the arguments have been seen previously. The
arguments to the previously memoized function, json-fodder-database-prepare-results,
hashed to the same value with each call; thus, sorted-snippets would return the same
results despite the underlying lambda function (key) differing between calls.

Because Pliny databases can now be created locally from a JSON flat file, I don't
believe it is worth the effort to determine how to memoize this properly.  This
has bit us before, for different reasons, in se:f90c95ee.