Fallback if no matching full-stmt in pick-general.
Previously, in pick-general, if we could not find a full statement
when the :full-stmt flag was set to true, a hard error would be thrown.

We now do not throw a hard error and offer non-matching ASTs as a fallback.
In general, this situation only occurs after a bad crossover, and the candidate
will be later filtered from the population.