Fix quotations in the -search macro.
Previously, the search macro would evaluate symbols instead of the value
of the symbol itself.  For instance, if we passed *analyze-mutations* to
search as the collect-mutation-stats parameter, mutation statistics
would be collected as the symbol *analyze-mutations* is non-nil.  We now
evaluate the underlying value of *analyze-mutations* (and other -search
parameters).