change dummy-asts restart to nullify-asts
This restart will simply return a nil list when a mutation error is
signaled during a call to `update-asts'.  Clients which currently try
`ignore-failed-mutation' restarts may want to consider also trying a
`nullify-asts' restart when no `ignore-failed-mutation' restart is
active.  (Note: the `find-restart' function may be used to test if a
named restart is currently active.)