Add searchable interface.
Create a 'searchable' interface for searching a backend for source snippets
similar to the given target.  This interface is a superclass of the existing
fodder database interface, which adds the ability to search for snippets of a
given AST class and other attributes.

The class hierarchy now looks as follows:

             +------------+
             | Searchable |
             +------------+
                   |
                   v
         +-----------------+
      +--| Fodder Database |----+
      |  +-----------------+    |
      |            |            |
      v            v            v
 +----------+ +----------+ +---------+
 | Pliny DB | | Mongo DB | | JSON DB |
 +----------+ +----------+ +---------+

This new hierarchy will support easily adding a RNN searchable backend to the
BED project.