Fix error in crossover with empty function bodies.
Previously, crossover of a software object involving an empty function body
would throw a hard error.  The new behaviour, recognizing valid crossover
points cannot be found, is to not perform a crossover at all and instead
return a copy of the first crossover parent.

I spent about 0.5 hours investigating what it would take to fully support
crossover of an empty function body; it does not appear that the current
implementation was designed with this in mind.  Indeed, the crossover
point selection routine ensures that only points "within" the function
body are selected.  Because a solution was not forthcoming immediately,
I decided to not to attempt to fully support crossover of empty function
bodies; if you would like, I can spend additional time on this.

This commit also fixes a small (but devastating) error in crossover.
If crossover points cannot be found, we must return a *copy* of the
first crossover parent.  Previously, we would return the parent itself;
as part of the crossover routine, after crossover, the returned object's
fitness is set to nil.  When we return the parent itself, instead of a copy,
an existing member of the population has it's fitness nil'd out.  This
leads to a fatal error when performing lexicase selection (due to unequal
fitness vectors in the population).