remove shuffle from json database initialization
This is done because the shuffle of such a large list can have a huge
performance impact on startup time for runs using the database.

This shuffle was added to ensure that snippets from the same program are
not located next to each other in the database.  This is because range
queries returning more than 1 snippet will take a continuous range from
the database.  If this will be a problem then the database should be
shuffled *on disk* before being loaded into lisp.