Change to computation of dyn mutation probs.
When computing dynamic mutation probabilities, for each mutation type,
we find the percentage of mutations which yielded better, same, worse,
or dead candidates.  We multiply each percentage by a bias, for
instance, *better-bias* and take the arithmatic mean of the result,
yielding a value, v.  Higher values of v indicate the mutation yielded
better candidates, while lower values indicate the converse.  The
maximum value of v is 1.25 (all better candidates), while the minimum
value of v is 0.00 (all dead candidates).

Previously, we used v as the new dynamic mutation probability.  This is
poor for a number of reasons.  For instance, if a single mutation of a
given mutation type was performed in the last checkpoint period and this
mutation yielded a better candidate, the dynamic probability of this
mutation would be 125% of all mutations, per the existing formula.

v is now used a weight to adjust the previous mutation probability, as
was the original intent.