expand-assignment may now be targeted.
Use the `bad-stmts' method to retrieve a list of `bad' statement ASTs.
This may be overridden in the caller (e.g. clang-w-binary) to return a
list of ASTs which are diff-inducing or some other notion of
`bad'. Select a `CompoundAssignOperator' AST from this list of ASTs for
expansion.  If an AST of the appropriate class cannot be found, expand
the search to all statement ASTs.

I'm including this for now, however moving forward I think the better
solution moving forward is that described in the

    TODO Should mutations fall back from bad-stmts to all-stmts

task of the NOTES file in this repository.  I'm adding a comment to this
effect in the code because long term this is the wrong place for logic
(on switching from targeted to un-targetted) which should be universal.

ok eschulte@ ip:VET-TA2