Restart in file-to-string to handle encoding.
In case of error, allow the caller to specify a different encoding and
try again. The encoding can also be passed directly if it's known.

The default encoding is UTF-8, which has the advantage that it can
detect errors if you try to read e.g. a Latin-1 file. Whereas reading a
UTF-8 file as Latin-1 will just produce junk characters.

ok eschulte@ IP:BRASS