Improve `fitness-better-p` for non-numeric fitness
Previously, when using non-numeric fitness, each item
in fitness-a had to be better than the corresponding
item in fitness-b according to *fitness-predicate*.
If fitness-a was better than fitness-b overall, but
had one item worse than fitness-b, `fitness-better-p`
would return false.  In this case, the converse would
also return false; thus, `fitness-better-p` did not
define strict order.  To define a strict ordering, we
need to convert the fitnesses to a numeric value which
summarizes overall fitness.

This commit adds a new global, *fitness-scalar-fn*,
which should be bound to a function which takes a
fitness value and converts it to a scalar.  It also
defines a default function for this purpose.