Prevent error in expand-arithmatic-op.
In C++ code, clang-mutate does not recognize member variables as
unbound, which can lead to arithmatic ops with no unbound vals. To avoid
crashing on this, get the variable from the first child (as was already
done for the += case) instead of from :unbound-vals.

This is also more correct if the LHS is a struct field
access (e.g. foo.bar++).