Remove failsafe from inward crossover.
This commit ensures that `prepare-inward-snippet` will not
enter infinite recursion without relying on a special failsafe.
This is done by replacing `get-enclosing-block` with
`get-parent` in select places within this method; the degenerate
infinite recursive case would be entered when the enclosing
block was not the parent of the given AST.  This change will not
affect crossover for the vast majority of cases.

This slightly changes the behaviour of crossover involving
switch-case statements by removing special logic which
was previously required for successful crossover.  This logic
is no longer needed, as the new solution is more general.