Instrumenting projects with multiple entry-objs
Some projects have multiple files with main functions. In this
case, the entry-obj argument to instrument may be a list. For
each object being instrumented, set its entry-obj to either
itself if it's in the entry-obj list or to the first entry-obj
in the entry-obj list. IP:BRASS

ok newston@ IP:BRASS