Initialize instrumentation before main.
C++ in particular can run code before main via static initializers. If
we instrument that code, it will try to write to __bi_mut_log_file
before it is opened, causing a crash.

Instead of initializing during main, use the "constructor" attribute to
run an initialization function on startup, before anything else. This is
non-standard, but gcc and clang both support it.