Compose mutations to generate new named mutations
This defines a new macro, `compose-mutations', which composes a list of
any number of mutations into a new single named mutation.  The new
mutation has a targeter which applies the targeters of each composed
mutation and a picker which applies the pickers of each composed
mutation.  This also defines `build-op' for the new mutation to apply
the composed mutations.

This commit also introduces another related function which should be
implemented at some point which will instead sequence the application of
mutations instead of composing their targeters and pickers.

(compose-mutations combo-m (m1 m2 m3)) ;; =>
   1. Collects targets for m1, m2, and m3
   2. Picks targets for m1, m2, and m3
   3. Mutates object using the three picked targets

(sequence-mutations combo-m (m1 m2 m3)) ;; =>
   1. Collects targets for m1, picks a target, and mutates object
   2. Collects targets for m2, picks a target, and mutates object
   3. Collects targets for m3, picks a target, and mutates object