Remove asts slot and avoid storing genome.
The genome method now walks the ASTs to retrieve the source text.

(setf genome) is still supported. It temporarily stores a string in the
genome slot until update-asts is called, at which point the string is
discarded.

The asts method is now equivalent to all-ast-paths -- these can probably
be merged in the future.