Don't copy genome of clang objects.
The genome slot should only be set temporarily until update-asts
runs. After that we assemble the genome by walking the ASTs. But the
vagaries of the copying process can leave both slots set in the copy,
leading to strange bugs. We can avoid that by forcing the genome to NIL
in the copy.

IP:BRASS