Simplify instrumentation setup.
When instrumenting an object, insert setup code if it contains
main(). Otherwise, insert an extern definition for the log variable, on
the assumption that some other file in the project will have the setup
code.

When instrumenting a project, also call log-to-filename on all
entry-objs, in case main() is located in one of them.