Sort mutation ops for all clang mutations.
Always apply mutation operations from the end of the genome backward, to
avoid conflicts where one op invalidates the targets of
another. Anything else can leave the AST in an inconsistent state, so it
makes sense to do this in one place rather than repeating it in various
individual mutations.

This also fixes intermittent problems with compose-mutations.