Update clang ASTs immediately after setting genome.
We previously delayed running update-asts as long as possible, which was
useful when it needed to be called after each mutation. However, this
causes problems with applicative ASTs. We always want update-asts to run
before copying, to avoid additional calls and duplicated data
structures. But it's difficult to do this without potential race
conditions.

By running update-asts right away, we sidestep both problems.