Updates to improve support for projects/BugInjectr
1. Adding a generic `apply-mutations` method to SEL.
This supports the --just parameter to BugInjector.
2. Adding project-specific implementations of `apply-mutations`
and `apply-picked-mutations` which consider all files in the
project when a `current file` is not set.
3. Adding project-specific implementations of `clang-format`
and `clang-tidy` to execute on all files when a `current file`
is not set.
4. Updating the `functions` argument to `instrument`.  Each
function now requires two parameters: (1) the original software
object and (2) the AST to instrument.  This allows projects to
use `var-instrument` and other functions which require a clang
software object as a parameter.
5. Adding a `clang-project` class to SEL.  This class creates
a project from a clang compilation database.  This class also
consolidates `clang` software object-specific routines previously
in the (now more generic) project class. Finally, this class
implements generic methods (to-file, instrument, collect-traces)
for use by BugInjector.

IP:BugInjector