Handle multi-byte characters in clang genomes.
clang-mutate gives the begin/end offsets of ASTs as bytes in the
original encoding, but CL strings are indexed by characters. This
mismatch can lead to garbled source text when we build the ASTs.

Work around this problem by scanning the genome for multi-byte
characters, and mapping byte offsets to character indices. This assumes
that files are written UTF-8 before calling clang-mutate, which is true
in both SBCL and CCL.