Preserve :bin symbol when collecting trace input.
collect-trace will convert the symbol :bin in the program inputs to the
path of the compiled executable. But in the :input entry of the trace,
it's better to save the original inputs with the :bin symbol rather than
the actual pass. This makes it easier to re-run the same trace on a
different program, for example.