Fix a bug in clang scopes.
Scope processing is meant to stop when we reach the original AST, so we
don't include variables which haven't been declared yet. But the check
for this was incorrect -- it assumes that the enclosing full statement
of an AST is a direct child of its enclosing scope, which is not always
true.

A better check is to use ast-later-p, which sidesteps any nesting issues
and ensures that we never accidentally include future declarations.