Collect multiple traces in collect-trace.
In BI we typically collect traces by running a test script, which may
execute the instrumented program multiple times for a given input. This
leads to problems where SEL closes its ends of the pipe after the first
execution, leaving the second waiting for a reader. This not only misses
traces, but can leave zombie processes around.

To avoid this, open the pipe repeatedly until it times out, collecting
one trace for each successful open. This means that collect-trace now
returns a list of traces, and we can end up with more than one trace per
input.

Opening the pipe with a timeout would normally require nonblocking IO,
but there doesn't seem to be a good CL interface to that. Instead, spawn
a thread to do the open call, and abort it after the timeout. This is
more complicated than I'd like but seems to work well.