Refactor various forms of type names.
When instrumenting we write pointers like "*type" but when creating
decls we use "type*". Because of this, passing around types can be
ambiguous, and it's better to use clang-type where possible.

To that end, I added type-trace-string and type-from-trace-string to do
conversions. For the decl form, type-decl-string already exists, and we
don't currently need to convert in the other direction.