Ensure -I flags point to a valid directory.
When creating a clang project, we build the project
within a temporary directory.  As such, -I flags
generally point to directories within this temp area
which is destroyed after the build.

To ensure proper operation with clang-mutate after
the temp area has been destroyed, this commit replaces
the -I flags pointing the temporary build area with
the analogous persistent project directory.