Handshake when collecting traces to prevent race conditions.
SEL signals that it's ready to read traces by creating a file with a
given name. The instrumented program waits for that file to exist,
then signals that it's ready to write by deleting the file. After that
both processes can open the FIFO and send traces as normal.

When a test script runs the instrumented program multiple times, this
forces each subsequent invocation to wait until we've collected the
previous trace before it opens its end of the pipe.