Rearrange instrumentation code to fix compilation problems.
Including unistd.h causes problems in some programs: in particular,
coreutils has its own version of that file which can't be included
before config.h. Work around this by moving the include and the trace
setup code at the end of the genome, while keeping the definition of
__sel_trace_file at the beginning so other instrumentation code can
reference it.