Fix instrumentation for projs with `other-files`.
If the `other-files` field is set on a clang project, we
insert log setup code into these files.  Previously, we would
pass `t` as the env-name parameter to `initialize-tracing` from the
`clang-project` instrument method.  This fails as `t` is not
a pathname.

There existed some logic in the `clang` instrument method
to set `trace-env` to `*instrument-log-env-name*` if this
value was passed as `t`.  To ensure consistency between
`clang` and `clang-project` software objects, this logic has
been moved into the `initialize-tracing` function.  This ensures
that for both `clang` and `clang-project` software objects, calling
`instrument` with :trace-env as `t` will succeed and use
`*instrument-log-env-name*`.

Frankly, this seems convoluted to begin with, but existing
comments seem to indicate this is the desired approach.