Store traces in a trace database.
Rather than returning the traces directly, store them in a database
which is saved in (traces obj).

This change removes the :predicate argument to collect-traces (which
is not currently used by any callers). Since we no longer retrieve the
traces from the database at this point, we can't filter them.