Improved handling of temp file creation errors.
In `collect-traces`, do not throw a hard error if a temporary
file cannot be created.  Instead, continue to retry until
a timeout is reached.

Additionally, reduce the probability of a hard error being
thrown by reducing the number of `access` calls in the
instrumented binary thru adding a call to `sleep` in the
loop which checks if the handshake file has been created.