Compile with ASDF instead of buildapp
This commit adds defsystem forms for each executable to the relevant
.asd files.  It also simplifies the Makefiles cl.mk and Makefile to
remove all references to buildapp and build with ASDF instead.

This should make building executables more robust, it should allow for
the automatic static linking of C code, and it should allow for
building executables across more common lisp implementations.

See:
http://fare.tunes.org/files/asdf3/asdf3-2014.html
https://common-lisp.net/project/ecl/static/manual/re55.html
https://www.reddit.com/r/lisp/comments/5jbsle/how_to_build_an_executable_with_sbcl/

This commit also adds "repl" targets to the Makefile to support
launching the REPL from the command line.