Consistent error behavior in `phenome` method.
`clang` and `project` software objects now may throw a `phenome` error
when compilation fails, allowing for interchangeable use by clients.
This commit adds a `ignore-phenome-errors` macro for clients which
wish to preserve the existing behavior and return nil from phenome
when compilation fails.

Additional discussion of this change and the decisions for it can be
found at:

https://git.grammatech.com/synthesis/sel/commit/e58430aba2e8b0d158a0ff5e61acb6a57c6804a3#note_8619

This commit also removes \*compilation-timeout\* from SEL.  This was
added to timeout BED software objects which exercised compiler bugs,
leading to hanging.  This logic has been moved into BED.

Finally, this commit renames `compile-p` to `phenome-p` and promotes
it to work on all software objects.