Adding JAVA support to BugInjector.
This work was done by a third-party group, Fraunhofer USA.
This commit squashes (and cleans up) a number of changes made
by this group.  The original repository may be
found at https://github.com/GrammaTech/sel-fraunhofer;
branches to note are 'develop' and 'fraunhofer-2018.01.21'.

This code may be considered a work-in-progress.  Future
work should include adding support for writing JAVA
traces in a binary format, improving the genericity
of JAVA software objects, and adding support for in-memory
ASTs to JAVA.

Given tight deadlines, these will not be able to be completed prior to
this merge.  This represents a "best-effort" to get JAVA support over
the finish line and into SEL.