Allow `make check' target to pass on docker images
Fix the test/bin/example-* scripts to allow for
execution within the docker image.  Use Clozure CCL
to remain consistent in LISP implementation used for
tests and to avoid having to set SBCL_HOME.  Additionally,
allow building when USER_QUICK_LISP is not set.

Additionally, add the definition of the `test'
function to examples #3 and #4 to allow to successful
compilation.