Fix bug in tests invoking binary multiple times.
If the executable is terminated, some data may remain in the pipe,
and when re-opened to read the next trace, this data will be
parsed as the trace header, leading to an error.  We now print
the name of a temporary file to use for the current trace to
the the handshake file.  This ensures a unique fifo queue
will be used for each trace, and thus, there will not be stall data.