Fixup clang project creation on Arch Linux.
Previously, clang project creation would fail on Arch
linux due to a mismatch between compilation database
format on Ubuntu and Arch.  On Ubuntu, the file
field on the compilation database is a full,
absolute path, whereas on Arch, this is left as
only the filename.  This difference led to several
errors when creating and using projects.

This has been resolved by ensuring full, absolute paths
are used where required on both Arch and Ubuntu.