Fixes to project creation to support Apache.
No assumptions about the contents of the :directory
and :file components about a compilation database
entry created by bear may be made other
than the combined path is the absolute path
to the compiled file.  A file, /path/to/file.c
may be represented as any of:

file: /path/to/file.c
directory: /path/to/

file: file.c
directory: /path/to

file: to/file.c
directory: /path/

The previous iteration assumed that the compilation
database was formatted such that the directory
component always contained the absolute path up to
the file.  This was incorrect.  Merging the
file and directory components is always required
when performing path manipulations.