Remove use of 'snippet' outside fodder db context.
snippet->ast was renamed from-alist.
ast->snippet was renamed to-alist.

References to 'snippet' within outside of fodder database context
were renamed to limit the use of the term 'snippet' to
fodder database operations.  The file generate-helpers.lisp,
which contained 'snippet' references, was removed after
consultation with @nweston as it is obsolete.