Coq: update tokenizer, add code to generate exprs.
* Change tokenizer to create lists when parens are encountered in types
(rather than inserting symbols for them).
* Add code to create Coq AST expressions. Very liberal about making
assumptions that might have to be adjusted later: specifically, we've
fixed the values for some poorly documented fields based on a small
number of examples, and it's possible that new examples will require
changes in some of those fields.
* Update whitespace parser to replace inhibited whitespace chars with a
space. Update unit tests accordingly.
* Update so that inserting dummy statement runs Exec in SerAPI. If Exec
isn't run, some output may be delayed and included with the next query,
which could cause incorrect or unexpected behavior.