Only track Coq imports at top of file.
The rest have to be left inline since order can matter (e.g., when
importing notations, Vector and List notations are the same, so a
function may be interpreted as using the wrong one if both are
imported).

We still store imports that are at the top of the file separately so
that we can reload quickly. The imports field on Coq objects now only
includes imports from the top of the file.