Fixed several minor issues in DEFTEST, DEFSUITE
DEFSUITE was always activating the SEL TEST suite,
rather than the TEST suite for the intended package.

DEFTEST now will add a (declare (long-running t))
declaration when :long-running is specified.

DEFTEST now properly handles declarations and doc
strings in the body of the test being defined.

LONG-RUNNING-P function returns true if a test is flagged
as long-running.

(deftest (foo :long-running) () ...) is equivalent to
(deftest foo () (declare (long-running t)) ...)

If a test is declared as long-running, a warning will never
be output regarding that fact. It had been issuing warnings
even if the function was already flagged as long-running.