Fixed: Tests being included unexpectedly
The root suite was randomly getting assigned, as its
parent, a test suite from a different project. It was
packing up the value of stefil::*suite* which could be
left from a previous load or evaluation.
This fix addresses that and cleans up a couple other
macro-related issues.
With this fix (sel/test::test) runs only the SEL tests,
and (shaker/test::test) runs only the SHAKER tests,
as expected.