Fix path issues linking libs in csurf-asm phenome
When building the linker command for csurf-asm phenomes, use the
hard-coded path read from the log-file only if that path exists. If it
doesn't, use -lspec when the name starts with lib (e.g. -lc for libc),
or use -l :filename for a non-standard name.

This addresses issues where the csurf log file was created on a system
whose linked libraries are stored in a different location than those on
which the SEL software object is being used. Note that linking issues
will still occur if the linked libraries are not installed or not on the
path.