build-target->artifacts, plural and not in command
This changes the previously default behavior of appending the
build-target to the build-command for projects.  Now build targets
should be explicitly included in the build command.

This also changes the name of the build-target field to artifacts.

Software projects now build a list of artifacts (instead of a single
string project).