Fix bug in csurf-asm linking.
Fix a bug where linker flags weren't set correctly for libraries on a
different path from the log file whose names include a version (e.g.,
libc.so.6). Add test that will expose this if run on arch (or other OS
that stores libs in a different location than Ubuntu 16).